mod: import chain, keys, and forOwn from lodash directly
At the small expense of immediate readability (i.e.,
where did these functions come from), but the potential
gain of minimizing the end bundle size, prefer to import
functions directly from a library (especially a big
one like this).

Incidentally, we can do the same with `axios`, and in fact,
we can replace `axios` with the built-in `fetch`. That
will come in a later commit.